+++ src/extra_checks/ast/__init__.py
@@ -1,7 +1,8 @@
-from typing import Type
+from typing import Container, Type
 
 from django.db import models
 
+from ..check_id import CheckId
 from .ast import ModelAST
 from .exceptions import MissingASTError
 from .protocols import (
@@ -14,8 +15,9 @@ from .protocols import (
 
 def get_model_ast(
     model_cls: Type[models.Model],
+    meta_checks: Container[CheckId],
 ) -> ModelASTDisableCommentProtocol:
-    return ModelAST(model_cls)
+    return ModelAST(model_cls, meta_checks)
 
 
 __all__ = [